requirements.txt
python-dotenv
langchain
sentence-transformers
flask
cohere
streamlit
langchain_community
faiss-cpu
numpy
